Restructures package
